updated x64 lib

+++ hardware/core-video-capture/src-native/linux/support.cpp
@@ -15,8 +15,8 @@ extern "C" void emulate_cc_delete(void* p) {
   if (p != 0)
     free(p);
 }
-void* operator new  (unsigned len) __attribute__((alias("emulate_cc_new")));
-void* operator new[](unsigned len) __attribute__((alias("emulate_cc_new")));   
+void* operator new  (size_t len) __attribute__((alias("emulate_cc_new")));
+void* operator new[](size_t len) __attribute__((alias("emulate_cc_new")));   
 void  operator delete  (void* p)   __attribute__((alias("emulate_cc_delete")));
 void  operator delete[](void* p)   __attribute__((alias("emulate_cc_delete")));
 void* __cxa_pure_virtual = 0;
